+++ src/api/v1/views/teachers.py
@@ -2,36 +2,38 @@ from models.teachers_and_degree import Teacher
 from api.v1.views import teacher_bp
 from api.engine import db
 import bcrypt
-from flask import jsonify, request
+from flask import abort, jsonify, request
 from sqlalchemy.exc import NoResultFound, PendingRollbackError
-from models.base_model import BaseModel
 from datetime import date
-
+from flask_login import current_user, login_required
 BASE_URL = 'http://localhost:5000/api/v1'
 
 
 @teacher_bp.route('/teachers', methods=['GET'], strict_slashes=False)
+@login_required
 def teachers():
     """return all Teacher instance from the storage"""
+    if current_user.__tablename__ == 'students':
+        abort(403)
     new_obj = {}
     all_teachers = []
     teachers = db.get_all_object(Teacher)
     if teachers:
         for tchr in teachers:
             department = [
-                f'{BASE_URL}/departments/{dept.dept_code }' for dept in tchr.departments]
+                f'{BASE_URL}/departments/{dept.dept_code }' for dept in tchr.departments if tchr.departments]
             scored = [
-                f'{BASE_URL}/scores/{score.id }' for score in tchr.scored]
+                f'{BASE_URL}/scores/{score.id }' for score in tchr.scored if tchr.scored]
             degrees = [
                 f'{BASE_URL}/degrees/{degree.id }' for degree in tchr.degrees if degree]
             courses = [
-                f'{BASE_URL}/courses/{course.course_code }' for course in tchr.courses]
+                f'{BASE_URL}/courses/{course.course_code }' for course in tchr.courses if tchr.courses]
             assignments = [
-                f'{BASE_URL}/assignments/{assign.id }' for assign in tchr.assignments]
+                f'{BASE_URL}/assignments/{assign.id }' for assign in tchr.assignments if tchr.assignments]
             communications = [
-                f'{BASE_URL}/communications/{comm.id }' for comm in tchr.communications]
+                f'{BASE_URL}/communications/{comm.id }' for comm in tchr.communications if tchr.communications]
             materials = [
-                f'{BASE_URL}/materials/{material.id }' for material in tchr.materials]
+                f'{BASE_URL}/materials/{material.id }' for material in tchr.materials if tchr.materials]
 
             for k, v in tchr.to_json().items():
                 if k in ['id', 'first_name', 'last_name',  'email', 'citizenship', 'tel',
@@ -46,15 +48,22 @@ def teachers():
             new_obj['communications'] = communications
             new_obj['assignments'] = assignments
 
-            all_teachers.append(new_obj)
+            if current_user.__tablename__ == 'admins':
+                all_teachers.append(new_obj)
+            if current_user.__tablename__ == 'teachers':
+                if tchr.id == current_user.id:
+                    all_teachers.append(new_obj)
             new_obj = {}
         return jsonify({"teachers": all_teachers}), 200
     return jsonify([]), 200
 
 
 @teacher_bp.route('/teachers/<int:id>', methods=['GET'], strict_slashes=False)
+@login_required
 def single_teacher(id):
     """return single Teacher instance from the storage"""
+    if current_user.__tablename__ == 'students':
+        abort(403)
     new_obj = {}
     teacher = db.get_by_id(Teacher, id)
     if teacher:
@@ -71,17 +80,31 @@ def single_teacher(id):
         materials = [
             f'{BASE_URL}/materials/{material.id }' for material in teacher.materials]
 
-        for k, v in teacher.to_json().items():
-            if k in ['id', 'first_name', 'last_name',  'email', 'citizenship', 'tel',
-                     'dob', 'staff_member', 'last_login', 'created_at', 'updated_at']:
-                new_obj[k] = v
+        if current_user.__tablename__ == 'admins':
+            for k, v in teacher.to_json().items():
+                if k in ['id', 'first_name', 'last_name',  'email', 'citizenship', 'tel',
+                         'dob', 'staff_member', 'last_login', 'created_at', 'updated_at']:
+                    new_obj[k] = v
 
-        new_obj['courses'] = courses
-        new_obj['materials'] = materials
-        new_obj['scored'] = scored
-        new_obj['department'] = department
-        new_obj['communications'] = communications
-        new_obj['assignments'] = assignments
+            new_obj['courses'] = courses
+            new_obj['materials'] = materials
+            new_obj['scored'] = scored
+            new_obj['department'] = department
+            new_obj['communications'] = communications
+            new_obj['assignments'] = assignments
+        if current_user.__tablename__ == 'teachers':
+            if teacher.id == current_user.id:
+                for k, v in teacher.to_json().items():
+                    if k in ['id', 'first_name', 'last_name',  'email', 'citizenship', 'tel',
+                             'dob', 'staff_member', 'last_login', 'created_at', 'updated_at']:
+                        new_obj[k] = v
+
+                new_obj['courses'] = courses
+                new_obj['materials'] = materials
+                new_obj['scored'] = scored
+                new_obj['department'] = department
+                new_obj['communications'] = communications
+                new_obj['assignments'] = assignments
 
         return jsonify({"teacher": new_obj}), 200
     return jsonify([]), 200
@@ -89,8 +112,11 @@ def single_teacher(id):
 
 @teacher_bp.route('/teachers', methods=['POST'],
                   strict_slashes=False)
+@login_required
 def create_teacher():
     """function that handles creation endpoint for Teacher instance"""
+    if current_user.__tablename__ != 'admins':
+        abort(404)
     data = dict(request.form)
 
     try:
@@ -114,8 +140,11 @@ def create_teacher():
 
 @teacher_bp.route('/teachers/<int:id>', methods=['PUT'],
                   strict_slashes=False)
+@login_required
 def update_teacher(id):
     """ function that handles update endpoint for Teacher instance"""
+    if current_user.__tablename__ != 'admins':
+        abort(404)
     data = dict(request.form)
     if data.get('staff_member'):
         data['staff_member'] = True
@@ -130,9 +159,11 @@ def update_teacher(id):
 
 @teacher_bp.route('/teachers/<int:id>', methods=['DELETE'],
                   strict_slashes=False)
+@login_required
 def delete_teacher(id):
     """function for delete endpoint, it handles Teacher deletion"""
-
+    if current_user.__tablename__ != 'admins':
+        abort(404)
     try:
         db.delete(Teacher, id)
     except NoResultFound as e:
